requirements.txt
raganything>=1.2.9
lightrag-hku>=1.4.9
qdrant-client>=1.7.0
streamlit>=1.28.0
python-dotenv>=1.0.0
numpy>=1.24.0
tqdm>=4.65.0
requests>=2.31.0
nest-asyncio>=1.5.0
azure-identity>=1.15.0
openai>=1.12.0
networkx>=3.0
